pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snapnarrate"
version = "0.3.2"
description = "SnapNarrate game narrator using OpenAI vision and ElevenLabs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "mss>=9.0.1",
  "Pillow>=10.4.0",
  "keyboard>=0.13.5",
  "pystray>=0.19.5",
  "numpy>=2.2.0",
  "sounddevice>=0.5.1",
  "soundfile>=0.13.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
snapnarrate = "snap_narrate.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

